+++ imio_survey/queriers/ArcRESTQuerier/querier.py
@@ -24,7 +24,11 @@ class ArcRESTQuerier(IQuerier):
         searchZone = self.geosGeom2EsriGeom(geosGeometry)
         mapService =  MapService(url)
         #TODO build mapextent and imageDisplay and so... results are wrong without correct parameters
-        result =  mapService.Identify(searchZone, sr="31370", layers=layers,tolerance=0, mapExtent="0,0,300000,300000", imageDisplay=1024, returnGeometry=False)
+        #layers top (default) // visible // all
+        layers_param="all"
+        if layers is not None:
+            layers_param = layers_param + ":" + layers
+        result =  mapService.Identify(searchZone, sr="31370", layers=layers_param,tolerance=0, mapExtent="0,0,300000,300000", imageDisplay=1024, returnGeometry=False)
         clean_results = []
         if result:
             for feat in result.results.features: